refactor(camel): refactor PropertiesTooltippedName and assign more specific id to Tooltip
