+++ packages/hawtio/src/util/jolokia.test.ts
@@ -1,17 +1,6 @@
-import { escapeHtmlId, escapeTags, operationToString } from './jolokia'
+import { operationToString } from './jolokia'
 
 describe('jolokia', () => {
-  test('escapeTags', () => {
-    expect(escapeTags('domain-name')).toEqual('domain-name')
-    expect(escapeTags('<domain-name>')).toEqual('&lt;domain-name&gt;')
-    expect(escapeTags('"SampleContext"')).toEqual('"SampleContext"')
-  })
-
-  test('escapeHtmlId', () => {
-    expect(escapeHtmlId('java.lang')).toEqual('java.lang')
-    expect(escapeHtmlId('Camel Contexts')).toEqual('CamelContexts')
-  })
-
   test('operationToString', () => {
     expect(operationToString('getCamelId', [])).toEqual('getCamelId()')
     expect(
